#1 bugs fixed & refactored
