+++ app/Http/Controllers/AuthenticationController.php
@@ -1,19 +1,17 @@
 <?php
 
-declare(strict_types=1);
-
 namespace App\Http\Controllers;
 
 use App\Http\Requests\LoginRequest;
 use App\Http\Requests\RegisterRequest;
-use App\Services\Interfaces\LoginServiceInterface;
-use App\Services\Interfaces\RegisterServiceInterface;
+use App\Services\Authentication\UserLoggerInterface;
+use App\Services\Authentication\UserRegisterInterface;
 use Illuminate\Http\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 
 class AuthenticationController extends Controller
 {
-    public function login(LoginRequest $request, LoginServiceInterface $service): JsonResponse
+    public function login(LoginRequest $request, UserLoggerInterface $service): JsonResponse
     {
         $token = $service->login($request->validated());
 
@@ -22,7 +20,7 @@ class AuthenticationController extends Controller
         ], Response::HTTP_OK);
     }
 
-    public function register(RegisterRequest $request, RegisterServiceInterface $service): JsonResponse
+    public function register(RegisterRequest $request, UserRegisterInterface $service): JsonResponse
     {
         $service->register($request->validated());
 
@@ -31,3 +29,4 @@ class AuthenticationController extends Controller
         ], Response::HTTP_OK);
     }
 }
+
