+++ app/Http/Requests/RegisterRequest.php
@@ -1,8 +1,8 @@
 <?php
 
-declare(strict_types=1);
 
 namespace App\Http\Requests;
+
 
 class RegisterRequest extends ApiRequest
 {
