+++ app/Services/Authentication/TestService.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace App\Services\Authentication;
+
+
+class TestService implements UserRegisterInterface
+{
+
+    public function register(array $credentials)
+    {
+        // TODO: Implement register() method.
+    }
+}
