+++ app/Services/Authentication/UserRegisterInterface.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Services\Interfaces;
+namespace App\Services\Authentication;
 
-interface RegisterServiceInterface
+interface UserRegisterInterface
 {
     public function register(array $credentials);
 }
